requirements.txt
blinker==1.9.0
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
efficientnet_pytorch==0.7.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.59.0
fsspec==2025.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpointer==2.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2025.2
PyYAML==6.0.2
scipy==1.16.1
setuptools==78.1.1
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tzdata==2025.2
ultralytics==8.3.178
ultralytics-thop==2.0.15
Werkzeug==3.1.3
wheel==0.45.1
flask_login==0.6.3